+++ components/ContactForm/const.js
@@ -1,0 +1,15 @@
+const initialFormState = {
+  name: "",
+  phone: "",
+  email: "",
+  message: "",
+};
+
+const initialErrorsState = {
+  name: false,
+  phone: false,
+  email: false,
+  message: false,
+};
+
+export { initialErrorsState, initialFormState };
